not finished
Added neccesary functionality to the bossmanger to make the scripts/actions/behaviors be able to run independent of the runFrames from the Phaseaction struct. Created a parent class for all actions in order to facilitate managing couroutines. Still having isse with making A0 repeat after every action that is not A0