more updates
Changed the BulletSpawner Script to inherit from BossAction instead of MonoBehaviour that way it could be used in the update manager, hid selectionChance variable as it is currently not working,  Changed the ChangePhase method to a coroutine to add the same functionality to the endBehaviour as the startBehaviour,
Fixed an unintended bug that would not allow the last possible action behavior to be picked due to the addition of the repeatingBehaviour mechanic. Also renabled the testBoss in the scene.